requirements.txt
desktop_notifier==6.0.0
gspread==6.1.4
oauth2client==4.1.3
pyserial==3.5
